TEST : acceptance 어노테이션 생성
