--- src/test/java/com/example/moviereview/MoviereviewApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.moviereview;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MoviereviewApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
