+++ src/test/java/com/moviereview/common/annotation/AcceptanceTest.java
@@ -1,0 +1,22 @@
+package com.moviereview.common.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.boot.test.autoconfigure.data.mongo.AutoConfigureDataMongo;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.test.annotation.DirtiesContext;
+
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@SpringBootTest(
+    properties = "de.flapdoodle.mongodb.embedded.version=5.0.5",
+    webEnvironment = WebEnvironment.RANDOM_PORT
+)
+@AutoConfigureDataMongo
+@DirtiesContext
+public @interface AcceptanceTest {
+}
